charts/nginz: Add restrictive CORS headers to all origins and explicitly allowlist allowed origins
There is no reason not to have restrictive CORS on all endpoints; given
our API is currently only consumed by internal services that we know
apriori. We can loosen this restriction again in the future (e.g. when
we have something like OAuth
integration).

We also move to an explicit allowlist instead of trusting the entirety
of `external_env_domain`. Especially on-prem people might have many more
services running on their base domain and we as Wire have no control
over it. We do not want other services not in our control to be a vector
for e.g. XSS attacks. Hence we require the user of our charts to
explicitly allowlist all the expected origins.  We will update the docs
to guide people into making this update.